+++ spec/spec_helper.rb
@@ -13,6 +13,8 @@ ENV["RAILS_ENV"] = "test"
 require File.expand_path("dummy/config/environment", __dir__)
 require "anyway_config"
 
+Anyway::Settings.use_local_files = false
+
 Rails.application.eager_load!
 
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }